requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==3.1.7
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.12.4
psycopg2-binary==2.8.6
pycodestyle==2.8.0
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
tzdata==2022.1
